rename some main classes / methods
